Fix legacy integration tests
